+++ examples/lstm_multi.h
@@ -44,11 +44,11 @@ private:
                 activation(activation::method::tanh, 0.0, 1.0),
                 0.0, // dropout
                 0.01, // weight decay
-                OptimiserType::Adam,
+                OptimiserType::AdamW,
                 0.9,
                 false, 0, 0, 0, 0, 0, 0, 0)
         },
-        OutputLayerDetails(5, activation(activation::method::tanh, 0.0, 1.0), ErrorCalculation::type::mse, EvaluationConfig(), 0.01, OptimiserType::Adam, 0.9)
+        OutputLayerDetails(5, activation(activation::method::tanh, 0.0, 1.0), ErrorCalculation::type::mse, EvaluationConfig(), 0.01, OptimiserType::AdamW, 0.9)
       ),
       MultiOutputLayerDetails(
         {
@@ -57,7 +57,7 @@ private:
                 activation(activation::method::tanh, 0.0, 1.0),
                 0.0, // dropout
                 0.01, // weight decay
-                OptimiserType::Adam,
+                OptimiserType::AdamW,
                 0.9,
                 false, 0, 0, 0, 0, 0, 0, 0),
           LayerDetails(Layer::Architecture::FF,
@@ -65,12 +65,12 @@ private:
                 activation(activation::method::tanh, 0.0, 1.0),
                 0.0, // dropout
                 0.01, // weight decay
-                OptimiserType::Adam,
+                OptimiserType::AdamW,
                 0.9, 
                 false, 0, 0, 0, 0, 0, 0, 0)
         },
         // Second output: Tanh, Temperature 1.0
-        OutputLayerDetails(2, activation(activation::method::tanh, 0.0, 1.0), ErrorCalculation::type::mse, EvaluationConfig(), 0.01, OptimiserType::Adam, 0.9)
+        OutputLayerDetails(2, activation(activation::method::tanh, 0.0, 1.0), ErrorCalculation::type::mse, EvaluationConfig(), 0.01, OptimiserType::AdamW, 0.9)
         )
 
     };